% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff16FvCB.R
\name{FF16FvCB_Plant}
\alias{FF16FvCB_Cohort}
\alias{FF16FvCB_Parameters}
\alias{FF16FvCB_Patch}
\alias{FF16FvCB_Plant}
\alias{FF16FvCB_PlantPlus}
\alias{FF16FvCB_SCM}
\alias{FF16FvCB_Species}
\alias{FF16FvCB_StochasticPatch}
\alias{FF16FvCB_StochasticPatchRunner}
\alias{FF16FvCB_StochasticSpecies}
\title{Create a FF16FvCB Plant or Cohort}
\usage{
FF16FvCB_Plant(s = FF16FvCB_Strategy())

FF16FvCB_Cohort(s = FF16FvCB_Strategy())

FF16FvCB_Species(s = FF16FvCB_Strategy())

FF16FvCB_Parameters(...)

FF16FvCB_Patch(p)

FF16FvCB_SCM(p)

FF16FvCB_StochasticSpecies(s = FF16FvCB_Strategy())

FF16FvCB_StochasticPatch(p)

FF16FvCB_StochasticPatchRunner(p)

FF16FvCB_PlantPlus(s = FF16FvCB_Strategy())
}
\arguments{
\item{s}{A \code{\link{FF16FvCB_Strategy}} object}

\item{...}{Arguments!}

\item{p}{A \code{Parameters<FF16FvCB>} object}
}
\description{
Create a FF16FvCB Plant or Cohort
}
\examples{
pl <- FF16FvCB_Plant()
pl$height
}

