% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R
\name{FF16FvCB_Strategy}
\alias{FF16FvCB_Strategy}
\title{Strategy parameters}
\usage{
FF16FvCB_Strategy(..., values = list(...))
}
\arguments{
\item{..., values}{Values to initialise the struct with (either as
variadic arguments, or as a list, but not both).}
}
\description{
Strategy parameters based on FvCB photosynthesis model.
}
